requirements.txt
django>=3.2,<4.0
django-baton
#Wagtail
wagtail==2.16.2

#django-social-share
wagtailgmaps
wagalytics==1.3
requests
celery
#gunicorn
uWSGI==2.0.20
redis
django-dotenv
psycopg2-binary
django-storages

#Django Web App dependencies

django-social-auth
social-auth-app-django
sorl-thumbnail
django-widget-tweaks
kombu==5.2.2
#easy-thumbnails==2.8
django-thumbnails
#reportlab
django-qr-code



boto3
#System HOST
PyOpenSSL
python-dotenv
djangorestframework
#python-memcached

#Ecommerce
braintree
weasyprint

#Eduaction
django-braces
django-embed-video
django-memcache-status

#Lenguajes
django-parler
django-extensions
django-localflavor
django-rosetta
